+++ cps/src/main/java/com/meesho/cps/data/request/CatalogCPCDiscountSaveRequest.java
@@ -15,7 +15,7 @@ public class CatalogCPCDiscountSaveRequest {
 
     @NotNull
     @JsonProperty("catalog_id")
-    private Integer catalogId;
+    private Long catalogId;
 
     @NotNull
     @JsonProperty("discount")
